% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_ttest.R
\name{kfold_ttest}
\alias{kfold_ttest}
\title{Compute correlated t-statistic and p-value for k-fold cross-validated results}
\usage{
kfold_ttest(x, y, n, k)
}
\arguments{
\item{x}{\code{numeric} vector of values for model A}

\item{y}{\code{numeric} vector of values for model B}

\item{n}{\code{integer} denoting total sample size}

\item{k}{\code{integer} denoting number of folds used in k-fold}
}
\value{
object of class \code{data.frame}
}
\description{
Compute correlated t-statistic and p-value for k-fold cross-validated results
}
\references{
Nadeau, C., and Bengio, Y. Inference for the Generalization Error. Machine Learning 52, (2003).

Corani, G., Benavoli, A., Demsar, J., Mangili, F., and Zaffalon, M. Statistical comparison of classifiers through Bayesian hierarchical modelling. Machine Learning, 106, (2017).
}
\author{
Trent Henderson
}
