% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampled_t_test.R
\name{resampled.t.test}
\alias{resampled.t.test}
\title{Compute correlated t-statistic and p-value for resampled data}
\usage{
resampled.t.test(x, y, n = length(x), n1, n2)
}
\arguments{
\item{x}{\code{numeric} vector of values for model A}

\item{y}{\code{numeric} vector of values for model B}

\item{n}{\code{integer} denoting number of repeat samples. Defaults to \code{length(x)}}

\item{n1}{\code{integer} denoting train set size}

\item{n2}{\code{integer} denoting test set size}
}
\value{
object of class \code{data.frame}
}
\description{
Compute correlated t-statistic and p-value for resampled data
}
\references{
Nadeau, C., and Bengio, Y. Inference for the Generalization Error. Machine Learning 52, (2003).
}
\author{
Trent Henderson
}
